Refactor cli into modules for better readability and maintenance
