+++ python/fedml/cli/cli_modules/build.py
@@ -1,0 +1,258 @@
+import os
+import shutil
+from os.path import expanduser
+
+import click
+
+from fedml.computing.scheduler.comm_utils.platform_utils import platform_is_valid
+
+FEDML_MLOPS_BUILD_PRE_IGNORE_LIST = 'dist-packages,client-package.zip,server-package.zip,__pycache__,*.pyc,*.git'
+
+
+@click.command("build", help="Build packages for MLOps platform (open.fedml.ai)")
+@click.help_option("--help", "-h")
+@click.option(
+    "--platform",
+    "-pf",
+    type=str,
+    default="octopus",
+    help="The platform name at the MLOps platform (options: octopus, parrot, spider, beehive, falcon, launch).",
+)
+@click.option(
+    "--type",
+    "-t",
+    type=str,
+    default="client",
+    help="client or server? (value: client; server)",
+)
+@click.option(
+    "--source_folder", "-sf", type=str, default="./", help="the source code folder path"
+)
+@click.option(
+    "--entry_point",
+    "-ep",
+    type=str,
+    default="./",
+    help="the entry point of the source code",
+)
+@click.option(
+    "--config_folder", "-cf", type=str, default="./", help="the config folder path"
+)
+@click.option(
+    "--dest_folder",
+    "-df",
+    type=str,
+    default="./",
+    help="the destination package folder path",
+)
+@click.option(
+    "--ignore",
+    "-ig",
+    type=str,
+    default="",
+    help="the ignore list for copying files, the format is as follows: *.model,__pycache__,*.data*, ",
+)
+def mlops_build(platform, type, source_folder, entry_point, config_folder, dest_folder, ignore):
+    click.echo("Argument for type: " + type)
+    click.echo("Argument for source folder: " + source_folder)
+    click.echo("Argument for entry point: " + entry_point)
+    click.echo("Argument for config folder: " + config_folder)
+    click.echo("Argument for destination package folder: " + dest_folder)
+    click.echo("Argument for ignore lists: " + ignore)
+
+    if not platform_is_valid(platform):
+        return
+
+    if type == "client" or type == "server":
+        click.echo(
+            "Now, you are building the fedml packages which will be used in the MLOps "
+            "platform."
+        )
+        click.echo(
+            "The packages will be used for client training and server aggregation."
+        )
+        click.echo(
+            "When the building process is completed, you will find the packages in the directory as follows: "
+            + os.path.join(dest_folder, "dist-packages")
+            + "."
+        )
+        click.echo(
+            "Then you may upload the packages on the configuration page in the MLOps platform to start the "
+            "federated learning flow."
+        )
+        click.echo("Building...")
+    else:
+        click.echo("You should specify the type argument value as client or server.")
+        exit(-1)
+
+    home_dir = expanduser("~")
+    mlops_build_path = os.path.join(home_dir, "fedml-mlops-build")
+    try:
+        shutil.rmtree(mlops_build_path, ignore_errors=True)
+    except Exception as e:
+        pass
+
+    ignore_list = "{},{}".format(ignore, FEDML_MLOPS_BUILD_PRE_IGNORE_LIST)
+    pip_source_dir = os.path.dirname(__file__)
+    pip_source_dir = os.path.dirname(pip_source_dir)
+    pip_build_path = os.path.join(pip_source_dir, "computing", "scheduler", "build-package")
+    build_dir_ignore = "__pycache__,*.pyc,*.git"
+    build_dir_ignore_list = tuple(build_dir_ignore.split(','))
+    shutil.copytree(pip_build_path, mlops_build_path,
+                    ignore_dangling_symlinks=True, ignore=shutil.ignore_patterns(*build_dir_ignore_list))
+
+    if type == "client":
+        result = build_mlops_package(
+            ignore_list,
+            source_folder,
+            entry_point,
+            config_folder,
+            dest_folder,
+            mlops_build_path,
+            "fedml-client",
+            "client-package",
+            "${FEDSYS.CLIENT_INDEX}",
+        )
+        if result != 0:
+            exit(result)
+        click.echo("You have finished all building process. ")
+        click.echo(
+            "Now you may use "
+            + os.path.join(dest_folder, "dist-packages", "client-package.zip")
+            + " to start your federated "
+              "learning run."
+        )
+    elif type == "server":
+        result = build_mlops_package(
+            ignore_list,
+            source_folder,
+            entry_point,
+            config_folder,
+            dest_folder,
+            mlops_build_path,
+            "fedml-server",
+            "server-package",
+            "0",
+        )
+        if result != 0:
+            exit(result)
+
+        click.echo("You have finished all building process. ")
+        click.echo(
+            "Now you may use "
+            + os.path.join(dest_folder, "dist-packages", "server-package.zip")
+            + " to start your federated "
+              "learning run."
+        )
+
+
+def build_mlops_package(
+        ignore,
+        source_folder,
+        entry_point,
+        config_folder,
+        dest_folder,
+        mlops_build_path,
+        mlops_package_parent_dir,
+        mlops_package_name,
+        rank,
+):
+    if not os.path.exists(source_folder):
+        click.echo("source folder is not exist: " + source_folder)
+        return -1
+
+    if not os.path.exists(os.path.join(source_folder, entry_point)):
+        click.echo(
+            "entry file: "
+            + entry_point
+            + " is not exist in the source folder: "
+            + source_folder
+        )
+        return -1
+
+    if not os.path.exists(config_folder):
+        click.echo("config folder is not exist: " + source_folder)
+        return -1
+
+    mlops_src = source_folder
+    mlops_src_entry = entry_point
+    mlops_conf = config_folder
+    cur_dir = mlops_build_path
+    mlops_package_base_dir = os.path.join(
+        cur_dir, "mlops-core", mlops_package_parent_dir
+    )
+    package_dir = os.path.join(mlops_package_base_dir, mlops_package_name)
+    fedml_dir = os.path.join(package_dir, "fedml")
+    mlops_dest = fedml_dir
+    mlops_dest_conf = os.path.join(fedml_dir, "config")
+    mlops_pkg_conf = os.path.join(package_dir, "conf", "fedml.yaml")
+    mlops_dest_entry = os.path.join("fedml", mlops_src_entry)
+    mlops_package_file_name = mlops_package_name + ".zip"
+    dist_package_dir = os.path.join(dest_folder, "dist-packages")
+    dist_package_file = os.path.join(dist_package_dir, mlops_package_file_name)
+    ignore_list = tuple(ignore.split(','))
+
+    shutil.rmtree(mlops_dest_conf, ignore_errors=True)
+    shutil.rmtree(mlops_dest, ignore_errors=True)
+    try:
+        shutil.copytree(mlops_src, mlops_dest, copy_function=shutil.copy,
+                        ignore_dangling_symlinks=True, ignore=shutil.ignore_patterns(*ignore_list))
+    except Exception as e:
+        pass
+    try:
+        shutil.copytree(mlops_conf, mlops_dest_conf, copy_function=shutil.copy,
+                        ignore_dangling_symlinks=True, ignore=shutil.ignore_patterns(*ignore_list))
+    except Exception as e:
+        pass
+    try:
+        os.remove(os.path.join(mlops_dest_conf, "mqtt_config.yaml"))
+        os.remove(os.path.join(mlops_dest_conf, "s3_config.yaml"))
+    except Exception as e:
+        pass
+
+    mlops_pkg_conf_file = open(mlops_pkg_conf, mode="w")
+    mlops_pkg_conf_file.writelines(
+        [
+            "entry_config: \n",
+            "  entry_file: " + mlops_dest_entry + "\n",
+            "  conf_file: config/fedml_config.yaml\n",
+            "dynamic_args:\n",
+            "  rank: " + rank + "\n",
+            "  run_id: ${FEDSYS.RUN_ID}\n",
+            # "  data_cache_dir: ${FEDSYS.PRIVATE_LOCAL_DATA}\n",
+            # "  data_cache_dir: /fedml/fedml-package/fedml/data\n",
+            "  mqtt_config_path: /fedml/fedml_config/mqtt_config.yaml\n",
+            "  s3_config_path: /fedml/fedml_config/s3_config.yaml\n",
+            "  log_file_dir: /fedml/fedml-package/fedml/data\n",
+            "  log_server_url: ${FEDSYS.LOG_SERVER_URL}\n",
+            "  client_id_list: ${FEDSYS.CLIENT_ID_LIST}\n",
+            "  client_objects: ${FEDSYS.CLIENT_OBJECT_LIST}\n",
+            "  is_using_local_data: ${FEDSYS.IS_USING_LOCAL_DATA}\n",
+            "  synthetic_data_url: ${FEDSYS.SYNTHETIC_DATA_URL}\n",
+            "  client_num_in_total: ${FEDSYS.CLIENT_NUM}\n",
+        ]
+    )
+    mlops_pkg_conf_file.flush()
+    mlops_pkg_conf_file.close()
+
+    local_mlops_package = os.path.join(mlops_package_base_dir, mlops_package_file_name)
+    if os.path.exists(local_mlops_package):
+        os.remove(os.path.join(mlops_package_base_dir, mlops_package_file_name))
+    mlops_archive_name = os.path.join(mlops_package_base_dir, mlops_package_name)
+    shutil.make_archive(
+        mlops_archive_name,
+        "zip",
+        root_dir=mlops_package_base_dir,
+        base_dir=mlops_package_name,
+    )
+    if not os.path.exists(dist_package_dir):
+        os.makedirs(dist_package_dir, exist_ok=True)
+    if os.path.exists(dist_package_file) and not os.path.isdir(dist_package_file):
+        os.remove(dist_package_file)
+    mlops_archive_zip_file = mlops_archive_name + ".zip"
+    if os.path.exists(mlops_archive_zip_file):
+        shutil.move(mlops_archive_zip_file, dist_package_file)
+
+    shutil.rmtree(mlops_build_path, ignore_errors=True)
+
+    return 0
