+++ python/fedml/cli/cli_modules/login.py
@@ -1,0 +1,206 @@
+import os
+
+import click
+
+from fedml.computing.scheduler.comm_utils import sys_utils
+from fedml.computing.scheduler.master.docker_login import login_with_server_docker_mode
+from fedml.computing.scheduler.master.server_constants import ServerConstants
+from fedml.computing.scheduler.master.server_login import logout as server_logout
+from fedml.computing.scheduler.slave.client_constants import ClientConstants
+from fedml.computing.scheduler.slave.client_login import logout as client_logout
+from fedml.computing.scheduler.slave.docker_login import login_with_docker_mode
+
+name = "login"
+
+
+@click.command("login", help="Login to MLOps platform")
+@click.help_option("--help", "-h")
+@click.argument("userid", nargs=-1)
+@click.option(
+    "--version",
+    "-v",
+    type=str,
+    default="release",
+    help="login to which version of MLOps platform. It should be dev, test or release",
+)
+@click.option(
+    "--client", "-c", default=None, is_flag=True, help="login as the FedML client.",
+)
+@click.option(
+    "--server", "-s", default=None, is_flag=True, help="login as the FedML server.",
+)
+@click.option(
+    "--api_key", "-k", type=str, default="", help="user api key.",
+)
+@click.option(
+    "--local_server",
+    "-ls",
+    type=str,
+    default="127.0.0.1",
+    help="local server address.",
+)
+@click.option(
+    "--role",
+    "-r",
+    type=str,
+    default="edge_server",
+    help="run as the role (options: edge_server, cloud_agent, cloud_server, edge_simulator, gpu_master_server.",
+)
+@click.option(
+    "--runner_cmd",
+    "-rc",
+    type=str,
+    default="{}",
+    help="runner commands (options: request json for start run, stop run).",
+)
+@click.option(
+    "--device_id", "-id", type=str, default="0", help="device id.",
+)
+@click.option(
+    "--os_name", "-os", type=str, default="", help="os name.",
+)
+@click.option(
+    "--docker", "-d", default=None, is_flag=True, help="login with docker mode at the client agent.",
+)
+@click.option(
+    "--docker-rank", "-dr", default="1", help="docker client rank index (from 1 to n).",
+)
+def mlops_login(
+        userid, version, client, server,
+        api_key, local_server, role, runner_cmd, device_id, os_name,
+        docker, docker_rank
+):
+    print("\n Welcome to FedML.ai! \n Start to login the current device to the MLOps (https://open.fedml.ai)...\n")
+    if userid is None or len(userid) <= 0:
+        click.echo("Please specify your account id, usage: fedml login $your_account_id")
+        return
+    account_id = userid[0]
+    platform_url = "open.fedml.ai"
+    if version != "release":
+        platform_url = "open-{}.fedml.ai".format(version)
+
+    # Check user id.
+    if userid == "":
+        click.echo(
+            "Please provide your account id in the MLOps platform ({}).".format(
+                platform_url
+            )
+        )
+        return
+    # click.echo("client {}, server {}".format(client, server))
+    # Set client as default entity.
+    is_client = client
+    is_server = server
+    if client is None and server is None:
+        is_client = True
+
+    # Check api key
+    user_api_key = api_key
+    if api_key is None:
+        user_api_key = "NONE"
+
+    # Check docker mode.
+    is_docker = docker
+    if docker is None:
+        is_docker = False
+
+    if is_client is True:
+        if is_docker:
+            login_with_docker_mode(account_id, version, docker_rank)
+            return
+        pip_source_dir = os.path.dirname(__file__)
+        pip_source_dir = os.path.dirname(pip_source_dir)
+        login_cmd = os.path.join(pip_source_dir, "computing", "scheduler", "slave", "client_daemon.py")
+
+        client_logout()
+        sys_utils.cleanup_login_process(ClientConstants.LOCAL_HOME_RUNNER_DIR_NAME,
+                                        ClientConstants.LOCAL_RUNNER_INFO_DIR_NAME)
+        sys_utils.cleanup_all_fedml_client_learning_processes()
+        sys_utils.cleanup_all_fedml_client_login_processes("client_login.py")
+        sys_utils.cleanup_all_fedml_client_api_processes(kill_all=True)
+        try:
+            ClientConstants.login_role_list.index(role)
+        except ValueError as e:
+            role = ClientConstants.login_role_list[ClientConstants.LOGIN_MODE_CLIEN_INDEX]
+
+        login_pid = sys_utils.run_subprocess_open(
+            [
+                sys_utils.get_python_program(),
+                "-W",
+                "ignore",
+                login_cmd,
+                "-t",
+                "login",
+                "-u",
+                str(account_id),
+                "-v",
+                version,
+                "-ls",
+                local_server,
+                "-r",
+                role,
+                "-id",
+                device_id,
+                "-os",
+                os_name,
+                "-k",
+                user_api_key,
+                "-ngc",
+                "0"
+            ]
+        ).pid
+        sys_utils.save_login_process(ClientConstants.LOCAL_HOME_RUNNER_DIR_NAME,
+                                     ClientConstants.LOCAL_RUNNER_INFO_DIR_NAME, login_pid)
+    if is_server is True:
+        # Check login mode.
+        try:
+            ServerConstants.login_role_list.index(role)
+        except ValueError as e:
+            click.echo(
+                "Please specify login mode as follows ({}).".format(
+                    str(ServerConstants.login_role_list)
+                )
+            )
+            return
+
+        if is_docker:
+            login_with_server_docker_mode(account_id, version, docker_rank)
+            return
+
+        pip_source_dir = os.path.dirname(__file__)
+        pip_source_dir = os.path.dirname(pip_source_dir)
+        login_cmd = os.path.join(pip_source_dir, "computing", "scheduler", "master", "server_daemon.py")
+        server_logout()
+        sys_utils.cleanup_login_process(ServerConstants.LOCAL_HOME_RUNNER_DIR_NAME,
+                                        ServerConstants.LOCAL_RUNNER_INFO_DIR_NAME)
+        sys_utils.cleanup_all_fedml_server_learning_processes()
+        sys_utils.cleanup_all_fedml_server_login_processes("server_login.py")
+        sys_utils.cleanup_all_fedml_server_api_processes(kill_all=True)
+        login_pid = sys_utils.run_subprocess_open(
+            [
+                sys_utils.get_python_program(),
+                "-W",
+                "ignore",
+                login_cmd,
+                "-t",
+                "login",
+                "-u",
+                str(account_id),
+                "-v",
+                version,
+                "-ls",
+                local_server,
+                "-r",
+                role,
+                "-rc",
+                runner_cmd,
+                "-id",
+                device_id,
+                "-os",
+                os_name,
+                "-k",
+                user_api_key
+            ]
+        ).pid
+        sys_utils.save_login_process(ServerConstants.LOCAL_HOME_RUNNER_DIR_NAME,
+                                     ServerConstants.LOCAL_RUNNER_INFO_DIR_NAME, login_pid)
