+++ python/fedml/cli/cli_modules/model.py
@@ -1,0 +1,303 @@
+import json
+
+import click
+
+from fedml.computing.scheduler.model_scheduler.device_model_cards import FedMLModelCards
+
+
+@click.group("model")
+def model():
+    """
+    Deploy and infer models.
+    """
+    pass
+
+
+@model.command("create", help="Create local model repository.")
+@click.help_option("--help", "-h")
+@click.option(
+    "--name", "-n", type=str, help="model name.",
+)
+def create_model(name):
+    if FedMLModelCards.get_instance().create_model(name):
+        click.echo("Create model {} successfully.".format(name))
+    else:
+        click.echo("Failed to create model {}.".format(name))
+
+
+@model.command("delete", help="Delete local model repository.")
+@click.help_option("--help", "-h")
+@click.option(
+    "--name", "-n", type=str, help="model name.",
+)
+def delete_model(name):
+    if FedMLModelCards.get_instance().delete_model(name):
+        click.echo("Delete model {} successfully.".format(name))
+    else:
+        click.echo("Failed to delete model {}.".format(name))
+
+
+@model.command("add", help="Add file to local model repository.")
+@click.help_option("--help", "-h")
+@click.option(
+    "--name", "-n", type=str, help="model name.",
+)
+@click.option(
+    "--path", "-p", type=str, help="path for specific model.",
+)
+def add_model_files(name, path):
+    if FedMLModelCards.get_instance().add_model_files(name, path):
+        click.echo("Add file to model {} successfully.".format(name))
+    else:
+        click.echo("Failed to add file to model {}.".format(name))
+
+
+@model.command("remove", help="Remove file from local model repository.")
+@click.help_option("--help", "-h")
+@click.option(
+    "--name", "-n", type=str, help="model name.",
+)
+@click.option(
+    "--file", "-f", type=str, help="file name for specific model.",
+)
+def remove_model_files(name, file):
+    if FedMLModelCards.get_instance().remove_model_files(name, file):
+        click.echo("Remove file from model {} successfully.".format(name))
+    else:
+        click.echo("Failed to remove file from model {}.".format(name))
+
+
+@model.command("list", help="List model in the local model repository.")
+@click.help_option("--help", "-h")
+@click.option(
+    "--name", "-n", type=str, help="model name.",
+)
+def list_models(name):
+    models = FedMLModelCards.get_instance().list_models(name)
+    if len(models) <= 0:
+        click.echo("Model list is empty.")
+    else:
+        for model_item in models:
+            click.echo(model_item)
+        click.echo("List model {} successfully.".format(name))
+
+
+@model.command("list-remote", help="List models in the remote model repository.")
+@click.help_option("--help", "-h")
+@click.option(
+    "--name", "-n", type=str, help="model name.",
+)
+@click.option(
+    "--user", "-u", type=str, help="user id.",
+)
+@click.option(
+    "--api_key", "-k", type=str, help="user api key.",
+)
+@click.option(
+    "--version",
+    "-v",
+    type=str,
+    default="release",
+    help="interact with which version of ModelOps platform. It should be dev, test or release",
+)
+@click.option(
+    "--local_server",
+    "-ls",
+    type=str,
+    default="127.0.0.1",
+    help="local server address.",
+)
+def list_remote_models(name, user, api_key, version, local_server):
+    if user is None or api_key is None:
+        click.echo("You must provide arguments for User Id and Api Key (use -u and -k options).")
+        return
+    FedMLModelCards.get_instance().set_config_version(version)
+    model_query_result = FedMLModelCards.get_instance().list_models(name, user, api_key, local_server)
+    if model_query_result is None or model_query_result.model_list is None or len(model_query_result.model_list) <= 0:
+        click.echo("Model list is empty.")
+    else:
+        click.echo("Found {} models:".format(len(model_query_result.model_list)))
+        index = 1
+        for model_item in model_query_result.model_list:
+            model_item.show("{}. ".format(index))
+            index += 1
+        click.echo("List model {} successfully.".format(name))
+
+
+@model.command("package", help="Build local model repository as zip model package.")
+@click.help_option("--help", "-h")
+@click.option(
+    "--name", "-n", type=str, help="model name.",
+)
+def package_model(name):
+    model_zip = FedMLModelCards.get_instance().build_model(name)
+    if model_zip != "":
+        click.echo("Build model package {} successfully".format(name))
+        click.echo("The local package is located at {}.".format(model_zip))
+    else:
+        click.echo("Failed to build model {}.".format(name))
+
+
+@model.command("push", help="Push local model repository to ModelOps(open.fedml.ai).")
+@click.help_option("--help", "-h")
+@click.option(
+    "--name", "-n", type=str, help="model name.",
+)
+@click.option(
+    "--model_storage_url", "-s", type=str, help="model storage url.",
+)
+@click.option(
+    "--model_net_url", "-mn", type=str, help="model net url.",
+)
+@click.option(
+    "--user", "-u", type=str, help="user id.",
+)
+@click.option(
+    "--api_key", "-k", type=str, help="user api key.",
+)
+@click.option(
+    "--version",
+    "-v",
+    type=str,
+    default="release",
+    help="interact with which version of ModelOps platform. It should be dev, test or release",
+)
+@click.option(
+    "--local_server",
+    "-ls",
+    type=str,
+    default="127.0.0.1",
+    help="local server address.",
+)
+def push_model(name, model_storage_url, model_net_url, user, api_key, version, local_server):
+    if user is None or api_key is None:
+        click.echo("You must provide arguments for User Id and Api Key (use -u and -k options).")
+        return
+    FedMLModelCards.get_instance().set_config_version(version)
+    model_is_from_open = True if model_storage_url is not None and model_storage_url != "" else False
+    model_storage_url, model_zip = FedMLModelCards.get_instance().push_model(name, user, api_key,
+                                                                             model_storage_url=model_storage_url,
+                                                                             model_net_url=model_net_url,
+                                                                             local_server=local_server)
+    if model_is_from_open:
+        click.echo("Push model {} with model storage url {} successfully.".format(name, model_storage_url))
+    else:
+        if model_storage_url != "":
+            click.echo("Push model {} successfully".format(name))
+            click.echo("The remote model storage is located at {}".format(model_storage_url))
+            click.echo("The local model package is locate at .".format(model_zip))
+        else:
+            click.echo("Failed to push model {}.".format(name))
+
+
+@model.command("pull", help="Pull remote model(ModelOps) to local model repository.")
+@click.help_option("--help", "-h")
+@click.option(
+    "--name", "-n", type=str, help="model name.",
+)
+@click.option(
+    "--user", "-u", type=str, help="user id.",
+)
+@click.option(
+    "--api_key", "-k", type=str, help="user api key.",
+)
+@click.option(
+    "--version",
+    "-v",
+    type=str,
+    default="release",
+    help="interact with which version of ModelOps platform. It should be dev, test or release",
+)
+@click.option(
+    "--local_server",
+    "-ls",
+    type=str,
+    default="127.0.0.1",
+    help="local server address.",
+)
+def pull_model(name, user, api_key, version, local_server):
+    if user is None or api_key is None:
+        click.echo("You must provide arguments for User Id and Api Key (use -u and -k options).")
+        return
+    FedMLModelCards.get_instance().set_config_version(version)
+    if FedMLModelCards.get_instance().pull_model(name, user, api_key, local_server):
+        click.echo("Pull model {} successfully.".format(name))
+    else:
+        click.echo("Failed to pull model {}.".format(name))
+
+
+@model.command("deploy",
+               help="Deploy specific model to ModelOps platform(open.fedml.ai) or just for local debugging deployment.")
+@click.help_option("--help", "-h")
+@click.option(
+    "--name", "-n", type=str, help="model name.",
+)
+@click.option(
+    "--on_premise", "-p", default=None, is_flag=True, help="all devices are from on-premise.",
+)
+@click.option(
+    "--cloud", "-c", default=None, is_flag=True, help="all devices are from fedml cloud.",
+)
+@click.option(
+    "--devices", "-d", type=str, help="device list, format: [1,2,3]. The first id is master device.",
+)
+@click.option(
+    "--user", "-u", type=str, help="user id or api key.",
+)
+@click.option(
+    "--api_key", "-k", type=str, help="user api key.",
+)
+@click.option(
+    "--params", "-pa", type=str, default="", help="serving parameters.",
+)
+@click.option(
+    "--version",
+    "-v",
+    type=str,
+    default="release",
+    help="interact with which version of ModelOps platform. It should be dev, test or release",
+)
+@click.option(
+    "--local_server",
+    "-ls",
+    type=str,
+    default="127.0.0.1",
+    help="local server address.",
+)
+@click.option(
+    "--use_local_deployment", "-ld", default=None, is_flag=True,
+    help="deploy local model repository by sending MQTT message(just use for debugging).",
+)
+def deploy_model(name, on_premise, cloud, devices, user, api_key, params, version,
+                 local_server, use_local_deployment):
+    if user is None or api_key is None:
+        click.echo("You must provide arguments for User Id and Api Key (use -u and -k options).")
+        return
+
+    is_cloud = cloud
+    is_on_premise = on_premise
+    if cloud is None and on_premise is None:
+        is_on_premise = True
+    if is_cloud and is_on_premise:
+        is_cloud = False
+
+    is_local_dev = use_local_deployment
+    if use_local_deployment is None:
+        is_local_dev = False
+
+    if is_on_premise:
+        device_type = "md.on_premise_device"
+    else:
+        device_type = "md.fedml_cloud_device"
+    FedMLModelCards.get_instance().set_config_version(version)
+
+    params_dict = {}
+    if is_local_dev:
+        params_dict = json.loads(params)  # load config from Cli
+
+    if FedMLModelCards.get_instance().deploy_model(name, device_type, devices, user, api_key,
+                                                   params_dict, use_local_deployment,
+                                                   local_server):
+        click.echo("Deploy model {} successfully.".format(name))
+    else:
+        click.echo("Failed to deploy model {}.".format(name))
